Move redis_v2 into redis
