+++ controllers/redisshard_controller.go
@@ -24,9 +24,9 @@ import (
 	basereconciler "github.com/3scale-ops/basereconciler/reconciler"
 	saasv1alpha1 "github.com/3scale/saas-operator/api/v1alpha1"
 	"github.com/3scale/saas-operator/pkg/generators/redisshard"
-	"github.com/3scale/saas-operator/pkg/redis_v2/client"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
-	"github.com/3scale/saas-operator/pkg/redis_v2/sharded"
+	"github.com/3scale/saas-operator/pkg/redis/client"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
+	"github.com/3scale/saas-operator/pkg/redis/sharded"
 	"github.com/3scale/saas-operator/pkg/util"
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
