+++ controllers/sentinel_controller.go
@@ -25,10 +25,10 @@ import (
 	saasv1alpha1 "github.com/3scale/saas-operator/api/v1alpha1"
 	"github.com/3scale/saas-operator/pkg/generators/sentinel"
 	"github.com/3scale/saas-operator/pkg/reconcilers/threads"
-	"github.com/3scale/saas-operator/pkg/redis_v2/events"
-	"github.com/3scale/saas-operator/pkg/redis_v2/metrics"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
-	"github.com/3scale/saas-operator/pkg/redis_v2/sharded"
+	"github.com/3scale/saas-operator/pkg/redis/events"
+	"github.com/3scale/saas-operator/pkg/redis/metrics"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
+	"github.com/3scale/saas-operator/pkg/redis/sharded"
 	"github.com/go-logr/logr"
 	grafanav1alpha1 "github.com/grafana-operator/grafana-operator/v4/api/integreatly/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
