+++ pkg/generators/twemproxyconfig/generator.go
@@ -10,8 +10,8 @@ import (
 	basereconciler_resources "github.com/3scale-ops/basereconciler/resources"
 	saasv1alpha1 "github.com/3scale/saas-operator/api/v1alpha1"
 	"github.com/3scale/saas-operator/pkg/generators"
-	"github.com/3scale/saas-operator/pkg/redis_v2/server"
-	"github.com/3scale/saas-operator/pkg/redis_v2/sharded"
+	"github.com/3scale/saas-operator/pkg/redis/server"
+	"github.com/3scale/saas-operator/pkg/redis/sharded"
 	"github.com/3scale/saas-operator/pkg/resource_builders/grafanadashboard"
 	"github.com/3scale/saas-operator/pkg/resource_builders/twemproxy"
 	"github.com/go-logr/logr"
