+++ pkg/redis/client/types_test.go
