+++ pkg/redis/events/watcher.go
@@ -5,11 +5,13 @@ import (
 	"fmt"
 
 	"github.com/3scale/saas-operator/pkg/reconcilers/threads"
-	"github.com/3scale/saas-operator/pkg/redis"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
+	"github.com/3scale/saas-operator/pkg/redis/sharded"
 	"github.com/go-logr/logr"
 	"github.com/prometheus/client_golang/prometheus"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/event"
+	"sigs.k8s.io/controller-runtime/pkg/metrics"
 )
 
 var (
@@ -66,26 +68,42 @@ var (
 
 func init() {
 	// Register custom metrics with the global prometheus registry
-	// metrics.Registry.MustRegister(switchMasterCount, failoverAbortNoGoodSlaveCount,
-	// 	sdownCount, sdownSentinelCount, sdownClearedCount, sdownClearedSentinelCount)
+	metrics.Registry.MustRegister(switchMasterCount, failoverAbortNoGoodSlaveCount,
+		sdownCount, sdownSentinelCount, sdownClearedCount, sdownClearedSentinelCount)
 }
 
 // SentinelEventWatcher implements RunnableThread
 var _ threads.RunnableThread = &SentinelEventWatcher{}
 
 type SentinelEventWatcher struct {
-	Instance      client.Object
-	SentinelURI   string
-	ExportMetrics bool
-	Topology      *redis.ShardedCluster
+	instance      client.Object
+	sentinelURI   string
+	exportMetrics bool
+	topology      *sharded.Cluster
 	eventsCh      chan event.GenericEvent
 	started       bool
 	cancel        context.CancelFunc
-	sentinel      *redis.SentinelServer
+	sentinel      *sharded.SentinelServer
+}
+
+func NewSentinelEventWatcher(sentinelURI string, instance client.Object, topology *sharded.Cluster,
+	metrics bool, pool *redis.ServerPool) (*SentinelEventWatcher, error) {
+	sentinel, err := sharded.NewSentinelServerFromPool(sentinelURI, nil, pool)
+	if err != nil {
+		return nil, err
+	}
+
+	return &SentinelEventWatcher{
+		instance:      instance,
+		sentinelURI:   sentinelURI,
+		exportMetrics: metrics,
+		topology:      topology,
+		sentinel:      sentinel,
+	}, nil
 }
 
 func (sew *SentinelEventWatcher) GetID() string {
-	return sew.SentinelURI
+	return sew.sentinelURI
 }
 
 // IsStarted returns whether the metrics gatherer is running or not
@@ -98,34 +116,27 @@ func (sew *SentinelEventWatcher) SetChannel(ch chan event.GenericEvent) {
 }
 
 func (sew *SentinelEventWatcher) Cleanup() error {
-	return sew.sentinel.CRUD.CloseClient()
+	return sew.sentinel.CloseClient()
 }
 
 // Start starts metrics gatherer for sentinel
 func (sew *SentinelEventWatcher) Start(parentCtx context.Context, l logr.Logger) error {
-	log := l.WithValues("sentinel", sew.SentinelURI)
+	log := l.WithValues("sentinel", sew.sentinelURI)
 	if sew.started {
 		log.Info("the event watcher is already running")
 		return nil
 	}
 
-	if sew.ExportMetrics {
+	if sew.exportMetrics {
 		// Initializes metrics with 0 value
 		sew.initCounters()
-	}
-
-	var err error
-	sew.sentinel, err = redis.NewSentinelServerFromConnectionString(sew.SentinelURI, sew.SentinelURI)
-	if err != nil {
-		log.Error(err, "cannot create SentinelServer")
-		return err
 	}
 
 	go func() {
 		var ctx context.Context
 		ctx, sew.cancel = context.WithCancel(parentCtx)
 
-		ch, closeWatch := sew.sentinel.CRUD.SentinelPSubscribe(ctx,
+		ch, closeWatch := sew.sentinel.SentinelPSubscribe(ctx,
 			`+switch-master`,
 			`-failover-abort-no-good-slave`,
 			`[+\-]sdown`,
@@ -139,7 +150,7 @@ func (sew *SentinelEventWatcher) Start(parentCtx context.Context, l logr.Logger)
 
 			case msg := <-ch:
 				log.V(1).Info("received event from sentinel", "event", msg.String())
-				sew.eventsCh <- event.GenericEvent{Object: sew.Instance}
+				sew.eventsCh <- event.GenericEvent{Object: sew.instance}
 				rem, err := NewRedisEventMessage(msg)
 				if err == nil {
 					log.V(3).Info("redis event message parsed",
@@ -149,7 +160,7 @@ func (sew *SentinelEventWatcher) Start(parentCtx context.Context, l logr.Logger)
 						"master-type", rem.master.role, "master-name", rem.master.name,
 						"master-ip", rem.master.ip, "master-port", rem.target.port,
 					)
-					if sew.ExportMetrics {
+					if sew.exportMetrics {
 						sew.metricsFromEvent(rem)
 					}
 				} else {
@@ -158,7 +169,6 @@ func (sew *SentinelEventWatcher) Start(parentCtx context.Context, l logr.Logger)
 
 			case <-ctx.Done():
 				log.Info("shutting down event watcher")
-				sew.sentinel.Cleanup(log)
 				sew.started = false
 				return
 			}
@@ -179,14 +189,14 @@ func (sew *SentinelEventWatcher) metricsFromEvent(rem RedisEventMessage) {
 	case "+switch-master":
 		switchMasterCount.With(
 			prometheus.Labels{
-				"sentinel": sew.SentinelURI, "shard": rem.master.name,
+				"sentinel": sew.sentinelURI, "shard": rem.master.name,
 				"redis_server": fmt.Sprintf("%s:%s", rem.master.ip, rem.master.port),
 			},
 		).Add(1)
 	case "-failover-abort-no-good-slave":
 		failoverAbortNoGoodSlaveCount.With(
 			prometheus.Labels{
-				"sentinel": sew.SentinelURI, "shard": rem.target.name,
+				"sentinel": sew.sentinelURI, "shard": rem.target.name,
 			},
 		).Add(1)
 	case "+sdown":
@@ -194,14 +204,14 @@ func (sew *SentinelEventWatcher) metricsFromEvent(rem RedisEventMessage) {
 		case "sentinel":
 			sdownSentinelCount.With(
 				prometheus.Labels{
-					"sentinel": sew.SentinelURI, "shard": rem.master.name,
+					"sentinel": sew.sentinelURI, "shard": rem.master.name,
 					"redis_server": fmt.Sprintf("%s:%s", rem.target.ip, rem.target.port),
 				},
 			).Add(1)
 		default:
 			sdownCount.With(
 				prometheus.Labels{
-					"sentinel": sew.SentinelURI, "shard": rem.master.name,
+					"sentinel": sew.sentinelURI, "shard": rem.master.name,
 					"redis_server": fmt.Sprintf("%s:%s", rem.target.ip, rem.target.port),
 				},
 			).Add(1)
@@ -211,14 +221,14 @@ func (sew *SentinelEventWatcher) metricsFromEvent(rem RedisEventMessage) {
 		case "sentinel":
 			sdownClearedSentinelCount.With(
 				prometheus.Labels{
-					"sentinel": sew.SentinelURI, "shard": rem.master.name,
+					"sentinel": sew.sentinelURI, "shard": rem.master.name,
 					"redis_server": fmt.Sprintf("%s:%s", rem.target.ip, rem.target.port),
 				},
 			).Add(1)
 		default:
 			sdownClearedCount.With(
 				prometheus.Labels{
-					"sentinel": sew.SentinelURI, "shard": rem.master.name,
+					"sentinel": sew.sentinelURI, "shard": rem.master.name,
 					"redis_server": fmt.Sprintf("%s:%s", rem.target.ip, rem.target.port),
 				},
 			).Add(1)
@@ -227,44 +237,44 @@ func (sew *SentinelEventWatcher) metricsFromEvent(rem RedisEventMessage) {
 }
 
 func (sew *SentinelEventWatcher) initCounters() {
-	if sew.Topology != nil {
+	if sew.topology != nil {
 
-		for _, shard := range *sew.Topology {
+		for _, shard := range sew.topology.Shards {
 			failoverAbortNoGoodSlaveCount.With(
 				prometheus.Labels{
-					"sentinel": sew.SentinelURI, "shard": shard.Name,
+					"sentinel": sew.sentinelURI, "shard": shard.Name,
 				},
 			).Add(0)
 
 			for _, server := range shard.Servers {
 				switchMasterCount.With(
 					prometheus.Labels{
-						"sentinel": sew.SentinelURI, "shard": shard.Name,
-						"redis_server": fmt.Sprintf("%s:%s", server.CRUD.IP, server.CRUD.Port),
+						"sentinel": sew.sentinelURI, "shard": shard.Name,
+						"redis_server": server.ID(),
 					},
 				).Add(0)
 				sdownSentinelCount.With(
 					prometheus.Labels{
-						"sentinel": sew.SentinelURI, "shard": shard.Name,
-						"redis_server": fmt.Sprintf("%s:%s", server.CRUD.IP, server.CRUD.Port),
+						"sentinel": sew.sentinelURI, "shard": shard.Name,
+						"redis_server": server.ID(),
 					},
 				).Add(0)
 				sdownCount.With(
 					prometheus.Labels{
-						"sentinel": sew.SentinelURI, "shard": shard.Name,
-						"redis_server": fmt.Sprintf("%s:%s", server.CRUD.IP, server.CRUD.Port),
+						"sentinel": sew.sentinelURI, "shard": shard.Name,
+						"redis_server": server.ID(),
 					},
 				).Add(0)
 				sdownClearedSentinelCount.With(
 					prometheus.Labels{
-						"sentinel": sew.SentinelURI, "shard": shard.Name,
-						"redis_server": fmt.Sprintf("%s:%s", server.CRUD.IP, server.CRUD.Port),
+						"sentinel": sew.sentinelURI, "shard": shard.Name,
+						"redis_server": server.ID(),
 					},
 				).Add(0)
 				sdownClearedCount.With(
 					prometheus.Labels{
-						"sentinel": sew.SentinelURI, "shard": shard.Name,
-						"redis_server": fmt.Sprintf("%s:%s", server.CRUD.IP, server.CRUD.Port),
+						"sentinel": sew.sentinelURI, "shard": shard.Name,
+						"redis_server": server.ID(),
 					},
 				).Add(0)
 			}
