+++ pkg/redis/server/server.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/3scale/saas-operator/pkg/redis_v2/client"
+	"github.com/3scale/saas-operator/pkg/redis/client"
 	"github.com/go-redis/redis/v8"
 )
