+++ pkg/redis/sharded/discover_test.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/3scale/saas-operator/pkg/redis_v2/client"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
+	"github.com/3scale/saas-operator/pkg/redis/client"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
 	"github.com/go-test/deep"
 )
