+++ pkg/redis/sharded/redis_server.go
@@ -1,8 +1,8 @@
 package sharded
 
 import (
-	"github.com/3scale/saas-operator/pkg/redis_v2/client"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
+	"github.com/3scale/saas-operator/pkg/redis/client"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
 )
 
 type RedisServer struct {
