+++ pkg/redis/sharded/redis_server_test.go
@@ -3,8 +3,8 @@ package sharded
 import (
 	"testing"
 
-	"github.com/3scale/saas-operator/pkg/redis_v2/client"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
+	"github.com/3scale/saas-operator/pkg/redis/client"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
 	"github.com/3scale/saas-operator/pkg/util"
 	"github.com/go-test/deep"
 )
