+++ pkg/redis/sharded/redis_shard.go
@@ -7,8 +7,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/3scale/saas-operator/pkg/redis_v2/client"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
+	"github.com/3scale/saas-operator/pkg/redis/client"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
 	"github.com/3scale/saas-operator/pkg/util"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
