+++ test/util/redis.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	saasv1alpha1 "github.com/3scale/saas-operator/api/v1alpha1"
-	redis "github.com/3scale/saas-operator/pkg/redis_v2/server"
+	redis "github.com/3scale/saas-operator/pkg/redis/server"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/rest"
 )
